Improve the room rolloff factor hanlding
Testing with Generic Software shows that the reverb room rolloff factor applies
to the currently selected distance model, not necessarily an inverse distance
model as described in the EFX docs. Though it should be noted, Generic Software
completely ignores AL_EFFECTSLOT_AUXILIARY_SEND_AUTO, never applies the cones
to the wet gain, and doesn't clamp to AL_MIX_GAIN and AL_MAX_GAIN for the wet
gains. It's unclear if the reverb or source room rolloff factors, or the
initial decay, should be imfluenced by the min or max gain properties.