Don't fail alcDevicePauseSOFT for disconnected devices
The device playback is stopped so the function doesn't technically fail. This
primarily just allows the device state to be flagged as paused for a subsequent
reset or reopen, which covers a potential issue where a device is disconnected,
the app pauses it due to something like an incoming call or being put into the
background, then successfully resets or reopens the device before playback
should resume. Without this, the device will (try to) start immediately upon
being reset or reopened, which would either fail (and disconnect again) or
start playing again prematurely before it can be stopped.