Use an atomic exchange to destruct the atomic unique_ptr
A relaxed load isn't required to synchronize with other threads. A thread could
exchange the held pointer, and a subsequent load is allowed to get either the
pre- or post-swap value (it's only required the value be "complete", not junk
or some mix of the old and new value). An exchange, however, is guaranteed to
synchronize with itself. If an atomic variable has the value 0, for example,
with one thread swapping in the value 1 and another the value 2, the two
threads are guaranteed to not both get 0 from the swap.