Add and use a function to get string_view lengths as an int
The .*s formatter needs the length of the subsequent string as a (signed) int,
but a string_view uses size_t for the length, which is an unsigned integer and
(potentially) larger. So the length has to be clamped to not overflow the int
and wrap to negative when cast.