Recognize the system device name for PipeWire and PulseAudio
This is for names like "alsa_output.pci-0000_08_00.4.analog-stereo". Names like
these are shared between PipeWire and PulseAudio APIs, offering a relatively
consistent system-level specifier (kind of like GUIDs on Windows). This doesn't
include the ALSA backend, though it's not commonly used outside of some power-
user setups (but it would be nice to find a way to identify the ALSA device
related to such a system device name).