Don't inline PhaseShifterT<S>::process
Being a template function, it's implicitly inline, but GCC doesn't want to
inline it and causes a warning. Since template functions are automatically
inline, the only way to undo the inline request is to never inline it
regardless of what the compiler would do if it was a normal function.