Add a LAF file player example
This is an object-based audio file format, allowing for more flexible channel
arrangements instead of the fixed quad, 5.1, 7.1, etc layouts (including with
height), along with dynamic movement of channels, where they'll be mixed in
real-time according to the device configuration.

Its primary drawback is the lack of any real compression, supporting only 8-,
16-, and 24-bit integer and 32-bit float PCM. Although it is capable of
skipping seconds-long periods of silence for each track, such space-saving
measures could be better served with proper compression like FLAC or Opus, the
latter of which at least is unlikely to work without some redesigns to the
format.

Regardless, this example player is still effective at showing how object-based
audio formats in general can be played with OpenAL, and can be adapted or
updated to other formats that may exist now or in the future.