Apply room rolloff factors even when Auxiliary Send Auto is off
This must've been based on a misunderstanding. The source and reverb room
rolloff factors are options to exaggerate *or replace* the automatic room
distance attenuation. Which means it needs to be able to apply without the
automatic room attenuation enabled. This may be why the room rolloff properties
still applied with AL_EFFECTSLOT_AUXILIARY_SEND_AUTO off in Generic Software,
assuming it wasn't a bug.