Don't send a debug message when the device disconnects
Debug messages should only be generated in response to calls by the app, rather
than asynchronous system events. It might be better to generate such a message
when an OpenAL function call detects the problem and deviates its behavior as a
result (e.g. starting a source and it becoming immediately AL_STOPPED because
the device is disconnected).