Only use noexcept in public headers on >= C++11
This resolves issues consuming openal-soft in projects that don't
explicitly specify a C++ standard of C++11 or newer.

This also handles the fact that __cplusplus is not set properly in MSVC
(always claiming to be 199711L) which prevented both the C++11 noexcept
and the C++17 noexcept annotations from taking effect.