Support finding CppWinRT through vcpkg
vcpkg provides a packaged version of CppWinRT with its own CMake target,
so if that's available then we can use it directly rather than trying to
install it from NuGet.

This resolves issues building for UWP within vcpkg, because vcpkg uses
Ninja internally and that doesn't pick up the props definition from the
NuGet package.