Don't use separate gains for the reverb early input coeffs
The input lines are both temporally and spatially separate, while the separate
gains were attempting to account for the time delay. However, as the lines are
also spatially separate, that caused sounds to be louder or quieter depending
on which lines responded most to a sound's direction (e.g. a sound from the
top-front-left direction would be on a line with no extra attenuation, while a
sound from the top-back-right direction would have the most extra attenuation),
which became more noticeable with shorter decay times causing more significant
attenuation from the time difference.

These input lines are also supposed to represent the first bounces of the room,
so it doesn't make much sense to attenuate them differently based on the room
decay time as these reflections all essentially had minimal room interaction.

Note that this will slightly increase the overall reverb volume, but individual
sounds may or may not end up subtly louder, and only to correct for them having
been erroneously attenuated due to their panning direction.