Enable -ftrivial-auto-var-init=pattern when available
This is essentially the same as the future C++26 feature to auto-initialize
local variables. It shouldn't cause much of a performance issue since most
variables are already initialized when declared, but the [[gnu::uninitialized]]
attribute can avoid it when needed, and the added safety isn't a bad thing for
the rest.

Note that MSVC seems to have a similar option with /RTCs, however in addition
to auto-initializing local variables, it also enables run-time checks for local
array over/underruns and stack pointer corruption, which adds more overhead
than I'm keen on. And there isn't mention of an attribute to avoid the auto-
initialization.