Don't autoselect the null backend
The null backend will always provide a dummy device, which outputs to nothing.
If none of the other backends initialize for some reason, falling back to the
null backend can make it seem like there's still a device even though it's not
accessing any system device. Excluding it from autoselection helps indicate
something is wrong.

The null backend will still be implicitly included if a drivers option is
specified with an ending comma (,) to indicate including all other unspecified
backends.