Remove the OtherIO backend
It's still just a skeleton, and it's not clear it will be very useful with
WASAPI exclusive mode being an option. But if it is, it would need significant
reworking to use a better COM threading setup like the WASAPI backend has, and
this commit will show what was there as a reference to rewrite it.