build: Add an option for whether to use libsndfile
OS distributions typically want to build a package with a predictable
set of options enabled, even if they are building in a non-minimal
environment where additional development headers/libraries are present.
For example,
<https://wiki.gentoo.org/wiki/Project:Quality_Assurance/Automagic_dependencies>
in Gentoo documentation explains this topic well.

With `-DALSOFT_SNDFILE=ON` (the default), libsndfile will still be used
if present, or not used if absent. With `-DALSOFT_SNDFILE=OFF`, it will
not be used, even if it happens to be present.

Signed-off-by: Simon McVittie <smcv@collabora.com>